Split routes into different files
